pyproject.toml
[project]
name = "litesubtitler"
version = "0.0.1"
description = "千言字幕助手"
authors = [
    {name = "Tobin",email = "924241281@qq.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydub (>=0.25.1,<0.26.0)",
    "openai (>=1.65.3,<2.0.0)",
    "retry (>=0.9.2,<0.10.0)",
    "pyqt6-tools (>=6.4.2.3.3,<7.0.0.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "pysrt (>=1.1.2,<2.0.0)",
]

[[tool.poetry.source]]
name = "aliyun"
url = "http://mirrors.aliyun.com/pypi/simple"
priority = "primary"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
